ISSUE #2495: Move console commands to the proper namespaces
