+++ phpunit.xml.dist
@@ -24,6 +24,7 @@
         <env name="APP_URL" value="http://localhost:8080/" force="true" />
         <env name="IMPORT_MODE" value="stravaApi" force="true" />
         <env name="AI_API_KEY" value="the-key" force="true" />
+        <env name="DEV_MODE_IS_ENABLED" value="0" force="true" />
     </php>
     <extensions>
         <bootstrap class="DAMA\DoctrineTestBundle\PHPUnit\PHPUnitExtension" />
