+++ src/Console/Cache/PruneRenderCacheConsoleCommand.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Console\Daemon;
+namespace App\Console\Cache;
 
 use App\Infrastructure\Cache\Render\RenderCache;
 use Symfony\Component\Console\Attribute\AsCommand;
