+++ src/Console/Strava/StravaViewWebhookSubscriptionConsoleCommand.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Console\Webhook;
+namespace App\Console\Strava;
 
 use App\Domain\Strava\Strava;
 use App\Infrastructure\Doctrine\Migrations\RequiresUpToDateDatabaseSchema;
@@ -11,15 +11,14 @@ use Monolog\Attribute\WithMonologChannel;
 use Psr\Log\LoggerInterface;
 use Symfony\Component\Console\Attribute\AsCommand;
 use Symfony\Component\Console\Command\Command;
-use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Console\Style\SymfonyStyle;
 
 #[WithMonologChannel('console-output')]
 #[RequiresUpToDateDatabaseSchema]
-#[AsCommand(name: 'app:strava:webhooks-unsubscribe', description: 'Delete a Strava webhook subscription')]
-final class StravaDeleteWebhookSubscriptionConsoleCommand extends Command
+#[AsCommand(name: 'app:strava:webhooks-view', description: 'View Strava webhook subscription(s)')]
+final class StravaViewWebhookSubscriptionConsoleCommand extends Command
 {
     public function __construct(
         private readonly Strava $strava,
@@ -28,24 +27,27 @@ final class StravaDeleteWebhookSubscriptionConsoleCommand extends Command
         parent::__construct();
     }
 
-    protected function configure(): void
-    {
-        $this->addArgument('subscriptionId', InputArgument::REQUIRED, 'The webhook subscription ID to delete');
-    }
-
     protected function execute(InputInterface $input, OutputInterface $output): int
     {
         $output = new SymfonyStyle($input, new LoggableConsoleOutput($output, $this->logger));
-        $subscriptionId = $input->getArgument('subscriptionId');
 
-        if (!$output->confirm(sprintf('Are you sure you want to delete subscription with ID %s?', $subscriptionId), false)) {
+        if (!$subscriptions = $this->strava->getWebhookSubscription()) {
+            $output->note('No webhook subscriptions found');
+            $output->note('Create a subscription with: docker compose exec app bin/console app:strava:webhooks-create');
+
             return Command::SUCCESS;
         }
 
-        $this->strava->deleteWebhookSubscription($subscriptionId);
-
-        $output->success('Webhook subscription deleted successfully!');
-        $output->comment('You will no longer receive automatic updates from Strava.');
+        $output->table(
+            headers: ['ID', 'Application ID', 'Callback URL', 'Created At', 'Updated At'],
+            rows: array_map(fn (array $subscription): array => [
+                $subscription['id'],
+                $subscription['application_id'],
+                $subscription['callback_url'],
+                $subscription['created_at'],
+                $subscription['updated_at'],
+            ], $subscriptions),
+        );
 
         return Command::SUCCESS;
     }
