+++ templates/html/finish-setup.html.twig
@@ -38,7 +38,7 @@
                     </div>
                     <div class="flex items-center gap-2 px-4 py-3 font-mono text-sm">
                         <span class="select-none text-primary">›</span>
-                        <code class="flex-1 overflow-x-auto whitespace-nowrap text-gray-800">docker compose exec app bin/console app:cron:run-strava-import</code>
+                        <code class="flex-1 overflow-x-auto whitespace-nowrap text-gray-800">docker compose exec app bin/console app:import:strava</code>
                     </div>
                 </div>
             </div>
